requirements.txt
pygame==2.6.1
Pillow==10.4.0
pywin32==308